중개수수료 계산 안되는 문제 해결

+++ app.py
@@ -100,7 +100,7 @@ def calculate_charge_value():
    input2 = request.args.get('input2')
 
 
-   expected_value = int(input1) * int(input2)
+   expected_value = float(input1) * float(input2)
 
    print(input1)
    print(input2)
